chore: move code, vsce still picks up wrong readme
