+++ .gitignore
@@ -2,7 +2,6 @@
 out
 dist
 node_modules
-.vscode-test/
 *.vsix
 
 # Test project build artifacts
@@ -11,8 +10,8 @@ test-rust-project/target/
 # Additional Rust build artifacts
 **/target/
 **/*.rs.bk
-packages/cargo_tools_vscode/extension/src/wasm/**/*
-packages/cargo_tools_vscode/extension/src/types/**/*
+vscode_extension/src/wasm/**/*
+vscode_extension/src/types/**/*
 
 # IDE and editor files
 **/.vscode/settings.json
