--- packages/cargo_tools_vscode/extension/.vscodeignore
@@ -1,10 +1,0 @@
-out/**
-node_modules/**
-src/**
-webpack.config.js
-**/tsconfig.json
-**/eslint.config.mjs
-**/*.map
-**/*.ts
-**/.vscode-test.*
-docs/**
