+++ vscode_extension/README.md
@@ -1,0 +1,37 @@
+# Cargo Tools
+
+![Build Status](https://github.com/NickelWenzel/cargo-tools/actions/workflows/ci.yml/badge.svg)
+![License](https://img.shields.io/badge/license-MIT-green)
+
+Cargo Tools is a Visual Studio Code extension that provides IDE-like features for Rust/Cargo development. It complements [rust-analyzer](https://marketplace.visualstudio.com/items?itemName=rust-lang.rust-analyzer) by adding project configuration controls, a workspace target browser, and cargo-make task management.
+
+## Features
+
+### Project Configuration
+
+![Project configuration](./media/cargo_tools_configuration.gif)
+
+### Project Outline
+
+![Project outline](./media/cargo_tools_project_outline.gif)
+
+### cargo-make Integration
+
+![cargo make](./media/cargo_tools_makefile_support.gif)
+
+## Documentation
+
+- [Online documentation](https://github.com/NickelWenzel/cargo-tools/packages/cargo_tools_vscode/extension/docs/README.md)
+
+## Requirements
+
+- Visual Studio Code 1.102.0 or higher
+- Rust toolchain with Cargo installed via [rustup](https://rustup.rs)
+
+## Inspiration
+
+This extension draws inspiration from [vscode-cmake-tools](https://marketplace.visualstudio.com/items?itemName=ms-vscode.cmake-tools).
+
+## License
+
+MIT — see [LICENSE](LICENSE) for details.
