+++ vscode_extension/webpack.config.js
@@ -1,0 +1,63 @@
+//@ts-check
+
+'use strict';
+
+const path = require('path');
+const WasmPackPlugin = require("@wasm-tool/wasm-pack-plugin");
+const CopyPlugin = require("copy-webpack-plugin");
+
+const extensionConfig = {
+  target: 'node20', // VS Code extensions run in a Node.js-context 📖 -> https://webpack.js.org/configuration/node/
+  mode: 'development', // this leaves the source code as close as possible to the original (when packaging we set this to 'production')
+
+  context: path.resolve(__dirname),
+  entry: './src/extension.ts', // the entry point of this extension, 📖 -> https://webpack.js.org/configuration/entry-context/
+  output: {
+    // the bundle is stored in the 'dist' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
+    path: path.resolve(__dirname, '../dist'),
+    filename: 'extension.js',
+    libraryTarget: 'commonjs2'
+  },
+  externals: {
+    vscode: 'commonjs vscode' // the vscode-module is created on-the-fly and must be excluded. Add other modules that cannot be webpack'ed, 📖 -> https://webpack.js.org/configuration/externals/
+    // modules added here also need to be added in the .vscodeignore file
+  },
+  resolve: {
+    // support reading TypeScript and JavaScript files, 📖 -> https://github.com/TypeStrong/ts-loader
+    extensions: ['.ts', '.js', '.wasm'],
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        exclude: /node_modules/,
+        use: 'ts-loader'
+      },
+      {
+        test: /\.wasm$/,
+        type: 'webassembly/async',
+      },
+    ]
+  },
+  plugins: [
+    new WasmPackPlugin({
+      crateDirectory: path.resolve(__dirname, "../packages/cargo_tools_vscode"),
+      outDir: path.resolve(__dirname, "./src/wasm"),
+      outName: "cargo_tools_vscode",
+      extraArgs: "--target nodejs",
+    }),
+    new CopyPlugin({
+      patterns: [
+        { from: "src/wasm/*.wasm", to: "[name][ext]" },
+      ],
+    }),
+  ],
+  devtool: 'source-map',
+  experiments: {
+    asyncWebAssembly: true, // enable wasm import() support
+  },
+  infrastructureLogging: {
+    level: "log", // enables logging required for problem matchers
+  },
+};
+module.exports = [extensionConfig];
